refactor: migrate jsBundleList to CSS modules
