+++ src/components/jsBundleList/styles.module.scss
@@ -1,0 +1,21 @@
+.Table {
+  display: block;
+  overflow: scroll;
+}
+
+.FileName {
+  font-size: 0.9em;
+  vertical-align: middle;
+  white-space: nowrap;
+}
+
+.Checksum {
+  vertical-align: middle;
+  width: 100%;
+}
+
+.ChecksumValue {
+  font-size: 0.75em;
+  white-space: nowrap;
+  color: var(--codeColor);
+}
